% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boxcox.R
\name{Boxcox}
\alias{Boxcox}
\title{Automatic Box Cox Transformation}
\usage{
Boxcox(Data, Method = c("MeanMedian", "Shapiro"))
}
\arguments{
\item{Data}{Numeric vector.}

\item{Method}{"MeanMedian" or "Shapiro". "MeanMedian" should be used if possible. "Shapiro" is an alternative if the "MeanMedian" result is not satisfactory.}
}
\value{
The numeric vector after the automatic Box Cox transformation.
}
\description{
Automatic Box Cox Transformation
}
