% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/data_cdf.R
\name{data_cdf}
\alias{data_cdf}
\title{Data Cumulative Distribution Function}
\usage{
data_cdf(x)
}
\arguments{
\item{x}{A vector of observations.}
}
\description{
Empirical Cumulative Distribution Function of Data.
}
\examples{
x = rdispl(n = 1e4, xmin = 10, alpha = 2.5, xmax = 1e5)
ecdf = data_cdf(x)
}
\keyword{cumulative}
\keyword{data}
\keyword{distribution}
\keyword{empirical}
\keyword{function}
\keyword{of}

