% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/displo.R
\name{displo}
\alias{displo}
\title{Discrete Powerlaw Object}
\usage{
displo(x, summary = TRUE)
}
\arguments{
\item{x}{A vector containing the observations.}

\item{summary}{Logical, whether print a summary with some information concerning data. By default is set to TRUE.}
}
\description{
This function allows to create a discrete powerlaw object to analyze.
}
\examples{
data(moby)
x = moby
o = displo(x)
}
\keyword{discrete}
\keyword{powerlaw}

