% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{Get a summary for a given data frame}
\usage{
eda(data, plot = TRUE)
}
\arguments{
\item{data}{Data frame or a single vector.}

\item{plot}{if TRUE histograms of numerical variables are plotted, TRUE by default.}
}
\value{
Metrics data frame
}
\description{
Get a summary for a given data frame
}
\examples{
eda(iris)
}
