% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastsum}
\alias{fastsum}
\title{Fast Sum implemented in Cpp.}
\usage{
fastsum(i, xmin, alpha)
}
\arguments{
\item{i}{An integer.}

\item{xmin}{An integer.}

\item{alpha}{A real number greater than 1.}
}
\description{
Cpp function which speed up the computation of the Hurwitz zeta function.
}

