% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.cat.R
\name{is.cat}
\alias{is.cat}
\title{Categorical check}
\usage{
is.cat(x, k = 2)
}
\arguments{
\item{x}{Vector}

\item{k}{Maximum unique elements such that is considered as categorical.}
}
\value{
TRUE or FALSE
}
\description{
Categorical check
}
\examples{
is.cat(round(runif(100)))
}
