% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/len.R
\name{len}
\alias{len}
\title{Length of a Vector}
\usage{
len(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Length function for lazy people.
}
\examples{
x = moby
n = len(x)
}
\keyword{length}
\keyword{vector}

