% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/MAE.R
\name{MAE}
\alias{MAE}
\title{Mean Absolute Error}
\usage{
MAE(x, y)
}
\arguments{
\item{x,y}{Two vectors of the same length.}
}
\description{
Mean Absolute Error.
}
\examples{
x = runif(10)
y = runif(10)
MAE(x,y)
}
\keyword{absolute}
\keyword{error}
\keyword{mean}

