% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/MAPE.R
\name{MAPE}
\alias{MAPE}
\title{Mean Absolute Percentage Error}
\usage{
MAPE(x, y)
}
\arguments{
\item{x,y}{Two vectors of the same length.}
}
\description{
Mean Absolute Percentage Error.
}
\examples{
x = runif(10)
y = runif(10)
MAPE(x,y)
}
\keyword{absolute}
\keyword{error}
\keyword{mean}
\keyword{percentage}

