% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMLambda.R
\name{MMLambda}
\alias{MMLambda}
\title{Mean Median Lambda}
\usage{
MMLambda(Data, lambda = c(-10, 10))
}
\arguments{
\item{Data}{Numeric vector.}

\item{lambda}{Optimization interval.}
}
\value{
The lambda and the added value such that the mean is equal to the median.
}
\description{
Mean Median Lambda
}
