\name{moby}
\docType{data}
\alias{moby}
\title{Moby Dick word count}
\description{
The frequency of occurrence of unique words in the novel Moby Dick by Herman Melville.
}
\usage{moby}
\format{A vector.}
\source{M. E. J. Newman, Power laws, Pareto distributions and Zipf's law. Contemporary Physics 46, 323 (2005)}
\keyword{datasets}
