% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/MPE.R
\name{MPE}
\alias{MPE}
\title{Mean Percentage Error}
\usage{
MPE(x, y)
}
\arguments{
\item{x,y}{Two vectors of the same length.}
}
\description{
Mean Percentage Error.
}
\examples{
x = runif(10)
y = runif(10)
MPE(x,y)
}
\keyword{error}
\keyword{mean}
\keyword{percentage}

