% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/MSE.R
\name{MSE}
\alias{MSE}
\title{Mean Squared Error}
\usage{
MSE(x, y)
}
\arguments{
\item{x,y}{Two vectors of the same length.}
}
\description{
Mean Squared Error.
}
\examples{
x = runif(10)
y = runif(10)
MSE(x,y)
}
\keyword{error}
\keyword{mean}
\keyword{squared}

