% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadim.R
\name{pcadim}
\alias{pcadim}
\title{Dimension objective PCA}
\usage{
pcadim(Data, Dim)
}
\arguments{
\item{Data}{Numeric or complex matrix whose SVD decomposition is to be computed. Logical matrices are coerced to numeric.}

\item{Dim}{Exact sublinear dimension required.}
}
\value{
Scores, transpose of the eigen vectors matrix and the approximated data with the sublinear dimension required.
}
\description{
Dimension objective PCA
}
