% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcavar.R
\name{pcavar}
\alias{pcavar}
\title{Variance objective PCA}
\usage{
pcavar(Data, Minvar = 99.99999)
}
\arguments{
\item{Data}{Numeric or complex matrix whose SVD decomposition is to be computed. Logical matrices are coerced to numeric.}

\item{Minvar}{Minimum variance to be explained.}
}
\value{
Scores, transpose of the eigen vectors matrix and the approximated data with the required explained variance.
}
\description{
Variance objective PCA
}
