% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotfit.R
\name{plotfit}
\alias{plotfit}
\title{Plot Fit of Discrete Powerlaw Distributions}
\usage{
plotfit(o, xmax = 1e+05)
}
\arguments{
\item{o}{A discrete powerlaw object.}

\item{xmax}{The maximum value to show.}
}
\description{
Plot fit of discrete powerlaw distributions.
}
\examples{
x = moby
o = displo(x)
getXmin(o)
plotfit(o)
}
\keyword{distributions}
\keyword{fit}
\keyword{plot}
\keyword{powerlaw}

