% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plothill.R
\name{plothill}
\alias{plothill}
\title{Hill plot}
\usage{
plothill(o, gxmin = 0, xmax = 1e+05)
}
\arguments{
\item{o}{A discrete powerlaw object.}

\item{gxmin}{Guess on the true value of the lower bound.}

\item{xmax}{Maximum value considered as candidate for the lower bound. Default is set to 1e5.}
}
\description{
Hill plot for discrete power law distributions.
}
\examples{
x = moby
o = displo(x)
plothill(o)
}
\keyword{hill}
\keyword{plot}

