% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotmultifit.R
\name{plotmultifit}
\alias{plotmultifit}
\title{Plot Multiple Fit of Discrete Powerlaw Distributions}
\usage{
plotmultifit(o)
}
\arguments{
\item{o}{A discrete powerlaw object.}
}
\description{
Plot multiple fit of discrete powerlaw distributions.
}
\examples{
x = moby
o = displo(x)
getXmin(o)
plotmultifit(o)
}
\keyword{distributions}
\keyword{fit}
\keyword{plot}
\keyword{powerlaw}

