% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotols.R
\name{plotols}
\alias{plotols}
\title{Plot OLS Fit of Discrete Powerlaw Distributions}
\usage{
plotols(o)
}
\arguments{
\item{o}{A discrete powerlaw object.}
}
\description{
Plot OLS fit of discrete powerlaw distributions.
}
\examples{
x = moby
o = displo(x)
plotols(o)
}
\keyword{distributions}
\keyword{fit}
\keyword{plot}
\keyword{powerlaw}

