% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pmf.R
\name{pmf}
\alias{pmf}
\title{Probability Mass Function}
\usage{
pmf(x)
}
\arguments{
\item{x}{A vector of observations.}
}
\description{
Empirical Probability Mass Function.
}
\examples{
x = rdispl(n = 1e4, xmin = 10, alpha = 2.5, xmax = 1e5)
obs = pmf(x)$x
probs = pmf(x)$y
}
\keyword{empirical}
\keyword{function}
\keyword{mass}
\keyword{probability}

