% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/rdispl.R
\name{rdispl}
\alias{rdispl}
\title{Discrete Powerlaw Random Generator}
\usage{
rdispl(n, xmin, alpha, xmax = 1e+05)
}
\arguments{
\item{n}{Number of observations.}

\item{xmin}{The lower bound of the powerlaw distribution.}

\item{alpha}{The scaling parameter.}

\item{xmax}{The maximum value generated.}
}
\description{
Random generator of discrete power law distribution with parameters xmin and alpha.
}
\examples{
x = rdispl(n = 1e4, xmin = 10, alpha = 2.5, xmax = 1e5)
}
\keyword{discrete}
\keyword{generator}
\keyword{powerlaw}
\keyword{random}

