% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/RMSE.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Squared Error}
\usage{
RMSE(x, y)
}
\arguments{
\item{x,y}{Two vectors of the same length.}
}
\description{
Root Mean Squared Error.
}
\examples{
x = runif(10)
y = runif(10)
RMSE(x,y)
}
\keyword{error}
\keyword{mean}
\keyword{root}
\keyword{squared}

