% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHLambda.R
\name{SHLambda}
\alias{SHLambda}
\title{Shapiro Lambda}
\usage{
SHLambda(Data, lambda = c(-10, 10))
}
\arguments{
\item{Data}{Numeric vector.}

\item{lambda}{Optimization interval.}
}
\value{
The lambda and the added value such that the p value of the shapiro test for normality is maximum.
}
\description{
Shapiro Lambda
}
