% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/std.R
\name{std}
\alias{std}
\title{Unity-based Normalization}
\usage{
std(x)
}
\arguments{
\item{x}{A vector to normalize.}
}
\description{
Unity-based normalization of a vector.
}
\examples{
x = moby
z = std(x)
}
\keyword{distributions}
\keyword{fit}
\keyword{plot}
\keyword{powerlaw}

